gemini-live-genai-python-sdk/requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
httpx
